+++ examples/using-remark-copy-linked-files/src/templates/blog-post.js
@@ -1,5 +1,6 @@
 import React from "react"
 import Helmet from "react-helmet"
+import { graphql } from "gatsby"
 import get from "lodash/get"
 
 import Layout from "../components/layout"