+++ packages/gatsby/src/internal-plugins/query-runner/query-runner.js
@@ -152,6 +152,8 @@ ${formatErrorDetails(errorDetails)}`)
       },
     })
 
-    return
+    return result
   }
+
+  return result
 }