+++ packages/gatsby/src/utils/html-renderer-queue.js
@@ -1,43 +1,28 @@
-const Queue = require(`better-queue`)
 const convertHrtime = require(`convert-hrtime`)
 const Worker = require(`jest-worker`).default
-const physicalCpuCount = require(`physical-cpu-count`)
+const numWorkers = require(`physical-cpu-count`) || 1
+const { chunk } = require(`lodash`)
 
-const myWorker = new Worker(require.resolve(`./worker`), {
-  numWorkers: physicalCpuCount,
+const workerPool = new Worker(require.resolve(`./worker`), {
+  numWorkers,
 })
 
-module.exports = (htmlComponentRendererPath, pages, activity) =>
-  new Promise((resolve, reject) => {
-    const start = process.hrtime()
-    const queue = new Queue(
-      (path, callback) => {
-        myWorker
-          .renderHTML({ htmlComponentRendererPath, path })
-          .then(callback)
-          .catch(reject)
-      },
-      {
-        concurrent: 20,
-      }
+module.exports = async (htmlComponentRendererPath, pages, activity) => {
+  const start = process.hrtime()
+  const segments = chunk(pages, numWorkers)
+  const concurrency = 30
+
+  await Promise.all(
+    segments.map(paths =>
+      workerPool.renderHTML({ htmlComponentRendererPath, paths, concurrency })
     )
+  )
 
-    pages.forEach(page => {
-      queue.push(page)
-    })
-
-    queue.on(`task_finish`, () => {
-      const stats = queue.getStats()
-      if (activity) {
-        activity.setStatus(
-          `${stats.total}/${pages.length} ${(
-            stats.total / convertHrtime(process.hrtime(start)).seconds
-          ).toFixed(2)} pages/second`
-        )
-      }
-    })
-
-    queue.on(`drain`, () => {
-      resolve()
-    })
-  })
+  if (activity) {
+    activity.setStatus(
+      `${pages.length}/${pages.length} ${(
+        pages.length / convertHrtime(process.hrtime(start)).seconds
+      ).toFixed(2)} pages/second`
+    )
+  }
+}